commiting nearos

+++ app/Console/Commands/NearOSCommand.php
@@ -47,17 +47,23 @@ class NearOSCommand extends Command
 
         $departments = self::$department[$department];
 
-        Nearoutofstock::truncate();
-
         $store = $settings->store();
 
+        if($store->nearos_status === "backgroundprocess"){
+            $this->info("Near Os is already running");
+            return "already running";
+        }
+
         $settings->put('nearos_status', 'backgroundprocess');
+
+        Nearoutofstock::truncate();
 
         $day = $store->threshold_days;
         $supply_days = $store->supply_days;
         $threshold_day = $store->qty_to_buy_threshold;
-        $from = yesterdayDate();
+        $from =  date('Y-m-d', strtotime(' - '.$day.' days'));
         $to = todaysDate();
+
 
         Stock::where('status',"1")->where('reorder',1)->chunk(200,function($stocks) use(&$day,&$supply_days,&$threshold_day,&$from,&$to, &$departments, &$department){
             foreach ($stocks as $stock) {
@@ -79,21 +85,7 @@ class NearOSCommand extends Command
                 }
 
 
-                //for retail department
-                $_product =   Invoiceitembatch::select(
-                    'stock_id',
-                    DB::raw( 'SUM(quantity) as qty')
-                )
-                    ->where('department',"retail")
-                    ->where('stock_id',$stock->id)
-                    ->whereHas('stock',function($q){
-                        //$q->whereNull('group_id');
-                    })
-                    ->whereHas('invoice',function($q) use(&$from,$to){
-                        $q->whereBetween('invoice_date',[$from,$to]);
-                    })
-                    ->groupBy('stock_id')
-                    ->get();
+
                 if($_product->count() > 0) {
                     $p = $_product->first()->toArray();
                     $qty+=round(abs($p['qty']/$_product->first()->stock->box));
@@ -102,7 +94,6 @@ class NearOSCommand extends Command
 
                 $now_qty = $stock->totalBalance();
 
-
                 $po = Purchaseitem::where('stock_id',$stock->id)->whereHas('purchase',function($q){
                     $q->where('status_id','6');
                 })
@@ -110,9 +101,11 @@ class NearOSCommand extends Command
                     ->limit(1)->first();
 
 
+                $last_supplier = $stock->stockBatches()->orderBy('id','DESC')->limit(1)->first();
+
+
                 if($thresholad_score > $now_qty){
                     $qty_to_buy = $qty * $threshold_day;
-                    $last_supplier = $stock->stockBatches()->orderBy('id','DESC')->get()->first();
                     //trigger insertion
                     $insert = [
                         'stock_id'=>$stock->id,
@@ -122,18 +115,17 @@ class NearOSCommand extends Command
                         'last_purchase_date'=>($po->po->date_completed ?? NULL),
                         'qty_to_buy'=> $qty_to_buy,
                         'current_sold'=>$qty,
-                        'is_grouped'=>($stock->group_id ? 1 : 0),
-                        'group_os_id'=>$stock->group_id,
+                        'is_grouped'=>($stock->stockgroup_id ? 1 : 0),
+                        'group_os_id'=>$stock->stockgroup_id,
                         'last_po_batch'=>($po->id ?? NULL),
                         'threshold_value'=> $thresholad_score,
                         'current_qty'=> $stock->totalBalance(),
-                        'last_supplier'=> !empty($last_supplier->supplier_id) ? $last_supplier->supplier_id : NULL
+                        'supplier_id'=> $last_supplier->supplier_id ?? NULL
                     ];
                     Nearoutofstock::create($insert);
                     continue;
                 }else if($now_qty < 2){
                     $qty_to_buy = $qty * $threshold_day;
-                    $last_supplier = $stock->stockBatches()->orderBy('id','DESC')->get()->first();
                     $insert = [
                         'stock_id'=>$stock->id,
                         'threshold_type'=>"NORMAL",
@@ -143,12 +135,13 @@ class NearOSCommand extends Command
                         'qty_to_buy'=> $qty_to_buy,
                         'current_sold'=>$qty,
                         'last_po_batch'=>($po->id ?? NULL),
-                        'is_grouped'=>($stock->group_id ? 1 : 0),
-                        'group_os_id'=>$stock->group_id,
+                        'is_grouped'=>($stock->stockgroup_id ? 1 : 0),
+                        'group_os_id'=>$stock->stockgroup_id,
                         'threshold_value'=> $thresholad_score,
                         'current_qty'=> $stock->totalBalance(),
-                        'last_supplier'=> !empty($last_supplier->supplier_id) ? $last_supplier->supplier_id : NULL
+                        'supplier_id'=>  $last_supplier->supplier_id ?? NULL
                     ];
+
                     Nearoutofstock::create($insert);
                     continue;
                 }else{
@@ -156,7 +149,6 @@ class NearOSCommand extends Command
                     //threshold_type to be NOT-NORMAL
                     // You add it has a quick fix to your logic
                     $qty_to_buy = $qty * $threshold_day;
-                    $last_supplier = $stock->stockBatches()->orderBy('id','DESC')->get()->first();
                     $insert = [
                         'stock_id'=>$stock->id,
                         'threshold_type'=>"NOT-NORMAL",
@@ -166,17 +158,19 @@ class NearOSCommand extends Command
                         'qty_to_buy'=> $qty_to_buy,
                         'current_sold'=>$qty,
                         'last_po_batch'=>($po->id ?? NULL),
-                        'is_grouped'=>($stock->group_id ? 1 : 0),
-                        'group_os_id'=>$stock->group_id,
+                        'is_grouped'=>($stock->stockgroup_id ? 1 : 0),
+                        'group_os_id'=>$stock->stockgroup_id,
                         'threshold_value'=> $thresholad_score,
                         'current_qty'=> $stock->totalBalance(),
-                        'last_supplier'=> !empty($last_supplier->supplier_id) ? $last_supplier->supplier_id : NULL
+                        'supplier_id'=>  $last_supplier->supplier_id ?? NULL
                     ];
                    Nearoutofstock::create($insert);
                     continue;
                 }
             }
         });
+
+
 
         $groups = Stockgroup::where('status',"1")->get();
         foreach($groups as $group) {
@@ -202,26 +196,6 @@ class NearOSCommand extends Command
                 }
             }
 
-            //for retail department
-            $_product = DB::table('invoiceitembatches as batch_item')
-                ->select(DB::raw('SUM(batch_item.quantity/st.box) as qty'))
-                ->join('stocks as st', 'st.id', '=', 'batch_item.stock_id')
-                ->join('stockgroups as group', 'group.id', '=', 'st.stockgroup_id')
-                ->join('invoices as invoices', 'invoices.id', '=', 'batch_item.invoice_id')
-                ->where('st.stockgroup_id', '=', $group->id)
-                ->where('st.status', '=', 1)
-                ->where('st.reorder', '=', 1)
-                ->where('batch_item.department', "retail")
-                ->whereBetween('invoices.invoice_date', [$from, $to])
-                ->groupBy('st.stockgroup_id')
-                ->get();
-
-            if ($_product->count() > 0) {
-                foreach ($_product as $_prod) {
-                    $qty += $_prod->qty;
-                    //$now_qty += $group->totalRetailBalance();
-                }
-            }
 
             $thresholad_score = round(abs(($qty / $day) * $supply_days));
 
@@ -231,7 +205,7 @@ class NearOSCommand extends Command
                 $qty_to_buy = $qty * $threshold_day;
                 $insert = [
                     'stock_id' => NULL,
-                    'group_id' => $group->id,
+                    'stockgroup_id' => $group->id,
                     'last_po_batch' =>($poItem->id ?? NULL),
                     'threshold_type' => "THRESHOLD",
                     'os_type' => 'GROUP',
@@ -248,7 +222,7 @@ class NearOSCommand extends Command
                 $qty_to_buy = $qty * $threshold_day;
                 $insert = [
                     'stock_id' => NULL,
-                    'group_id' => $group->id,
+                    'stockgroup_id' => $group->id,
                     'threshold_type' => "NORMAL",
                     'os_type' => 'GROUP',
                     'last_po_batch' =>($poItem->id ?? NULL),
